[CHANGE] Move the Subversion chaining functionality into a dedicated subclass.          This preserves the original --stop-on-copy basic behavior in its own class.            * This keeps the base SvnAdapter class uncomplicated.            * The base SvnAdapter can still be used if desired.            * I can foresee another, better Subversion importer built around svnsync              that will not require chaining.
