+++ lib/scm.rb
@@ -13,6 +13,7 @@ require 'lib/scm/diff'
 require 'lib/scm/adapters/abstract_adapter'
 require 'lib/scm/adapters/cvs_adapter'
 require 'lib/scm/adapters/svn_adapter'
+require 'lib/scm/adapters/svn_chain_adapter'
 require 'lib/scm/adapters/git_adapter'
 require 'lib/scm/adapters/hg_adapter'
 require 'lib/scm/adapters/bzr_adapter'
