+++ lib/scm/adapters/svn_adapter.rb
@@ -13,4 +13,3 @@ require 'lib/scm/adapters/svn/push'
 require 'lib/scm/adapters/svn/pull'
 require 'lib/scm/adapters/svn/head'
 require 'lib/scm/adapters/svn/misc'
-require 'lib/scm/adapters/svn/chain'
