+++ lib/scm/adapters/svn_chain/cat_file.rb
@@ -1,0 +1,8 @@
+module Scm::Adapters
+	class SvnChainAdapter < SvnAdapter
+		def cat(path, revision)
+			parent_svn(revision) ? parent_svn.cat(path, revision) : super(path, revision)
+		end
+	end
+end
+
